feat: [#997] input schema checked, take only exp output cols

+++ transform/app/transform/schemas/output/offer.py
@@ -5,7 +5,7 @@ offer_output_schema = {
     "best_access_right": "string",
     "description": "array<string>",
     "id": "string",
-    "iid": "int",
+    "iid": "bigint",
     "internal": "boolean",
     "open_access": "boolean",
     "popularity": "int",
